socket communication

+++ host/python/specan_ui/dos_attacker.py
@@ -5,39 +5,39 @@ import signal
 import time
 import subprocess
 from subprocess import Popen, PIPE
+from multiprocessing.connection import Listener
 
 #
 #TODO:
 #read from file/fifo/stdin
 #blocking read, after read execute comparison and start hackrf_transfer
-#
-
-
 global prev_data
 prev_data = 0
-
+address = ('localhost', 6000)
+listener = Listener(address, authkey=b'secret password')
+print('connecting...')
+conn = listener.accept()
+print ('connection accepted from', listener.last_accepted)
 
 class main():
-    #os.mkfifo("/tmp/freqs")
-    fifo = os.open('/tmp/ubertooth.fifo', os.O_RDONLY)
+    #fifo = open('/tmp/ubertooth.fifo', 'rb')
     pro = 0
     while True:
-        data = os.read(fifo, 4)
-        [xdata] = struct.unpack('f', data)
-        print(xdata)
-        #data = input()
-        #print(data)
-        if(int.from_bytes(data, sys.byteorder)):
-        #if (abs(int(data,2) - prev_data) > 1):
+        data = conn.recv()
+        # do something with msg
+        if data == 'close':
+            conn.close()
+            break
+
+        if (abs(data - prev_data) > 1):
             print(data)
-            #print('data:', int(data,2), 'prev:', prev_data)
-            prev_data = int(data,2)
-            frequency = int(data,2) * 1e6
-            arg1 = 'hackrf_transfer -f ' + str(frequency) + ' -a 1 -x 40 -t /dev/zero'
-            arg2 = ' -f ' + str(frequency) + ' -a 1 -x 40 -t /dev/zero'
+            prev_data = data
+            frequency = data * 1e6
+            arg1 = 'hackrf_transfer -f ' + str(frequency) + ' -a 1 -x 36 -t /dev/zero'
+            #arg2 = ' -f ' + str(frequency) + ' -a 1 -x 40 -t /dev/zero'
             if pro:
                 os.killpg(os.getpgid(pro.pid), signal.SIGTERM)
-                time.sleep(3.5)
+                time.sleep(0.125)
             pro = subprocess.Popen(arg1, stdout=subprocess.PIPE, shell=True, preexec_fn=os.setsid)
 
             #subprocess.run(arg1 , shell=True, check=True)
@@ -58,5 +58,9 @@ class main():
             #    print('frequency= ', frequency)
 
         else:
-            print('else')
+           print('else')
+
+    print('closing socket...')
+    listener.close()
+    print('socket closed')
     print("Done")
